requirements.txt

# Downgrade to a stable combination with Pydantic v1
langchain==0.1.20
langchain-google-vertexai==0.1.3
langgraph==0.0.48
pydantic<2

# Other dependencies
arxiv==2.1.0
python-dotenv==1.0.1
pandas==2.2.2
schedule==1.2.2
PyPDF2==3.0.1
pdfplumber==0.11.1
pillow==10.3.0
langchain-text-splitters==0.0.1
sentence-transformers==2.2.2
scikit-learn==1.5.0
tiktoken==0.7.0
numpy==1.26.4
pymupdf==1.24.1
reportlab==4.2.0
streamlit==1.31.0
backoff==2.2.1
markdown>=3.8.0
